refactor: API 로직 분리 (#9)
