+++ src/app/(providers)/used-goods/_components/UsedGoodsList.tsx
@@ -2,9 +2,9 @@
 
 import { useQuery } from '@tanstack/react-query';
 import UsedGoodsItem from './UsedGoodsItem';
-import { getQueryFunction, getQueryKey } from '../page';
 import styles from './usedGoodsList.module.scss';
 import { useQueryParam } from '@/hooks/useQueryParam';
+import { getQueryKey, getQueryFunction } from '@/apis/goods';
 
 const UsedGoodsList = () => {
   const { isSoldout, queryObject } = useQueryParam();
@@ -14,14 +14,13 @@ const UsedGoodsList = () => {
   });
 
   // TODO: Empty Component 만들기
-
   if (!data) return <div>상품이 없습니다.</div>;
 
   const ForSaleGoods = data.filter((goods) => !goods.sold_out);
   const SoldOutGoods = data.filter((goods) => goods.sold_out);
 
-  if (isSoldout && !SoldOutGoods.length) return <div>상품이 없습니다.</div>;
-  if (!isSoldout && !ForSaleGoods.length) return <div>상품이 없습니다.</div>;
+  if ((isSoldout && !SoldOutGoods.length) || (!isSoldout && !ForSaleGoods.length))
+    return <div>상품이 없습니다.</div>;
 
   return (
     <div className={styles.wrapper}>
